Refactor main logic into CommandProcessor
